requirements.txt
sentence_transformers
tensorboardX
bm25s
